+++ tests/cli/test_suite.py
@@ -41,7 +41,7 @@ from tests.cli.utils import assert_no_logging_messages_or_tracebacks
 from tests.render.renderer.v3.test_suite_profile_notebook_renderer import (
     EXPECTED_EXPECTATION_CONFIGURATIONS_ONBOARDING_DATA_ASSISTANT,
 )
-from tests.render.test_util import (
+from render.util import (
     find_code_in_notebook,
     load_notebook_from_path,
     run_notebook,
