+++ tests/render/test_OpsgenieRenderer.py
@@ -1,8 +1,13 @@
+import pytest
+
 from great_expectations.core.batch import BatchDefinition, IDDict
 from great_expectations.core.expectation_validation_result import (
     ExpectationSuiteValidationResult,
 )
 from great_expectations.render.renderer import OpsgenieRenderer
+
+# module level markers
+pytestmark = [pytest.mark.big]
 
 
 def test_OpsgenieRenderer_validation_results_success():
