+++ tests/render/test_inline_renderer.py
@@ -16,6 +16,9 @@ from great_expectations.render.exceptions import InlineRendererError
 from great_expectations.render.renderer.inline_renderer import InlineRenderer
 from great_expectations.render.renderer_configuration import MetaNotesFormat
 
+# module level markers
+pytestmark = [pytest.mark.unit]
+
 
 def clean_serialized_rendered_atomic_content_graphs(
     serialized_rendered_atomic_content: List[dict],
@@ -29,7 +32,6 @@ def clean_serialized_rendered_atomic_content_graphs(
     return serialized_rendered_atomic_content
 
 
-@pytest.mark.unit
 def test_inline_renderer_instantiation_error_message(
     basic_expectation_suite: ExpectationSuite,
 ):
@@ -42,7 +44,6 @@ def test_inline_renderer_instantiation_error_message(
     )
 
 
-@pytest.mark.unit
 @pytest.mark.parametrize(
     "expectation_configuration,fake_result,expected_serialized_expectation_validation_result_rendered_atomic_content",
     [
@@ -411,7 +412,6 @@ def test_inline_renderer_expectation_validation_result_serialization(
     )
 
 
-@pytest.mark.unit
 @pytest.mark.parametrize(
     "expectation_configuration,expected_serialized_expectation_configuration_rendered_atomic_content",
     [
