+++ tests/render/test_renderer.py
@@ -7,6 +7,7 @@ from great_expectations.core.expectation_validation_result import (
 from great_expectations.render.renderer.renderer import Renderer
 
 
+@pytest.mark.unit
 def test_render():
     # noinspection PyUnusedLocal
     with pytest.raises(NotImplementedError):
@@ -41,14 +42,12 @@ def test__find_evr_by_type(titanic_profiled_evrs_1):
     found_evr = Renderer()._find_evr_by_type(
         titanic_profiled_evrs_1.results, "expect_column_to_exist"
     )
-    print(found_evr)
     assert found_evr is None
 
     # TODO: _find_all_evrs_by_type should accept an ValidationResultSuite, not ValidationResultSuite.results
     found_evr = Renderer()._find_evr_by_type(
         titanic_profiled_evrs_1.results, "expect_column_distinct_values_to_be_in_set"
     )
-    print(found_evr)
     assert found_evr == ExpectationValidationResult(
         success=True,
         result={
@@ -82,14 +81,12 @@ def test__find_all_evrs_by_type(titanic_profiled_evrs_1):
     found_evrs = Renderer()._find_all_evrs_by_type(
         titanic_profiled_evrs_1.results, "expect_column_to_exist", column_=None
     )
-    print(found_evrs)
     assert found_evrs == []
 
     # TODO: _find_all_evrs_by_type should accept an ValidationResultSuite, not ValidationResultSuite.results
     found_evrs = Renderer()._find_all_evrs_by_type(
         titanic_profiled_evrs_1.results, "expect_column_to_exist", column_="SexCode"
     )
-    print(found_evrs)
     assert found_evrs == []
 
     # TODO: _find_all_evrs_by_type should accept an ValidationResultSuite, not ValidationResultSuite.results
@@ -98,7 +95,6 @@ def test__find_all_evrs_by_type(titanic_profiled_evrs_1):
         "expect_column_distinct_values_to_be_in_set",
         column_=None,
     )
-    print(found_evrs)
     assert len(found_evrs) == 4
 
     # TODO: _find_all_evrs_by_type should accept an ValidationResultSuite, not ValidationResultSuite.results
@@ -107,13 +103,11 @@ def test__find_all_evrs_by_type(titanic_profiled_evrs_1):
         "expect_column_distinct_values_to_be_in_set",
         column_="SexCode",
     )
-    print(found_evrs)
     assert len(found_evrs) == 1
 
 
 def test__get_column_list_from_evrs(titanic_profiled_evrs_1):
     column_list = Renderer()._get_column_list_from_evrs(titanic_profiled_evrs_1)
-    print(column_list)
     assert column_list == [
         "Unnamed: 0",
         "Name",
