+++ tests/render/test_styled_string_template.py
@@ -1,4 +1,9 @@
+import pytest
+
 from great_expectations.render.view import DefaultJinjaPageView
+
+# module level markers
+pytestmark = [pytest.mark.unit]
 
 
 def test_render_template():
